# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: manageCallAI API
  version: 0.1.0
  summary: Business-level telecom control plane API over FreeSWITCH
  description: |
    OpenAPI contract for the manageCallAI REST API.

    **MVP response envelope**: all implemented endpoints return `{ "data": ... }` for
    both single-resource and list responses. Pagination will be introduced later
    where required. Error responses return `{ "error": "<CODE>", "message": "<human text>",
    "request_id": "<id>" }` where `error` is a machine-readable RPC code (NOT_FOUND,
    INVALID_ARGUMENT, UNAUTHENTICATED, PERMISSION_DENIED, ALREADY_EXISTS,
    RESOURCE_EXHAUSTED, INTERNAL, UNAVAILABLE).

    **Auth model**:
    - Human-facing endpoints: `Authorization: Bearer <JWT>` (issued by `/auth/login`
      or `/auth/register`). JWT claims include `sub`, `tenant_id`, `email`, `role`.
    - Runtime/internal endpoints: `Authorization: Bearer <RUNTIME_API_TOKEN>` or
      header `x-managecallai-runtime-token: <token>`. These are called by
      FreeSWITCH and the Go ESL agent — not by end users.

    Only implemented endpoints are documented here. Future resources (outbound
    routes, prompts, CDRs, audit events, approvals, simulations) will be added
    when implemented.
servers:
  - url: https://api.example.com/api/v1
    description: Production placeholder
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /inbound-routes:
    get:
      tags:
        - InboundRoutes
      summary: List inbound routes
      responses:
        "200":
          description: Inbound route list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundRouteListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    post:
      tags:
        - InboundRoutes
      summary: Create inbound route
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInboundRouteRequest"
      responses:
        "201":
          description: Inbound route created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundRouteResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /inbound-routes/{routeId}:
    parameters:
      - $ref: "#/components/parameters/RouteIdParam"
    get:
      tags:
        - InboundRoutes
      summary: Get inbound route with versions
      responses:
        "200":
          description: Inbound route
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundRouteResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    patch:
      tags:
        - InboundRoutes
      summary: Update inbound route metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInboundRouteRequest"
      responses:
        "200":
          description: Inbound route updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundRouteResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /inbound-routes/{routeId}/versions:
    parameters:
      - $ref: "#/components/parameters/RouteIdParam"
    post:
      tags:
        - InboundRoutes
      summary: Create draft version
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - definition
              properties:
                definition:
                  type: object
      responses:
        "201":
          description: Draft version created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouteVersionResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /inbound-routes/{routeId}/versions/{versionId}/validate:
    parameters:
      - $ref: "#/components/parameters/RouteIdParam"
      - $ref: "#/components/parameters/VersionIdParam"
    post:
      tags:
        - InboundRoutes
      summary: Validate inbound route version
      responses:
        "200":
          description: Validation passed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResultResponse"
        "422":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResultResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /inbound-routes/{routeId}/versions/{versionId}/publish:
    parameters:
      - $ref: "#/components/parameters/RouteIdParam"
      - $ref: "#/components/parameters/VersionIdParam"
    post:
      tags:
        - InboundRoutes
      summary: Publish validated version
      responses:
        "200":
          description: Route published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundRouteResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /inbound-routes/{routeId}/rollback:
    parameters:
      - $ref: "#/components/parameters/RouteIdParam"
    post:
      tags:
        - InboundRoutes
      summary: Roll back to previous published version
      responses:
        "200":
          description: Route rolled back
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboundRouteResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT issued by /auth/login or /auth/register
    runtimeTokenBearer:
      type: http
      scheme: bearer
      description: RUNTIME_API_TOKEN as Bearer token (preferred)
    runtimeTokenHeader:
      type: apiKey
      in: header
      name: x-managecallai-runtime-token
      description: RUNTIME_API_TOKEN as custom header
  parameters:
...
    RouteIdParam:
      name: routeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    VersionIdParam:
      name: versionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          enum:
            - NOT_FOUND
            - INVALID_ARGUMENT
            - UNAUTHENTICATED
            - PERMISSION_DENIED
            - ALREADY_EXISTS
            - RESOURCE_EXHAUSTED
            - INTERNAL
            - UNAVAILABLE
        message:
          type: string
        request_id:
          type: string
      required:
        - error
        - message
        - request_id
...